+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use go embed in place of pkger.
 
+### Fixed
+
+- Ensure Spark CR release version label is updated when upgrading a cluster.
+
 ## [5.8.1] - 2021-07-22
 
 ### Fixed